### 2023-09-21 #5861 @a91quaini
Hi,

thank you for your email. I'd like to suggest Fabio Ashtar Telarico (
https://reviewers.joss.theoj.org/reviewers/2760) as reviewer, since he is
proficient in R and his areas of expertise are related to my field.

Kind regards,
Alberto Quaini

On Thu, 21 Sept 2023 at 05:26, The Open Journals editorial robot <
***@***.***> wrote:

> *Submitting author:* @a91quaini <https://github.com/a91quaini> (Alberto
> Quaini <http://orcid.org/0000-0002-1251-0599>)
> *Repository:* https://github.com/a91quaini/intrinsicFRP
> *Branch with paper.md* (empty if default branch):
> *Version:* v1.0.0
> *Editor:* Pending
> *Reviewers:* Pending
> *Managing EiC:* Olivia Guest
> Status
>
> [image: status]
> <https://joss.theoj.org/papers/c64de1bb6589df77811e1aa42a1871de>
>
> Status badge code:
>
>
> *Author instructions*
>
> Thanks for submitting your paper to JOSS @a91quaini
> <https://github.com/a91quaini>. *Currently, there isn't a JOSS editor
> assigned* to your paper.
>
> @a91quaini <https://github.com/a91quaini> if you have any suggestions for
> potential reviewers then please mention them here in this thread (without
> tagging them with an @). You can search the list of people
> <https://reviewers.joss.theoj.org/lookup> that have already agreed to
> review and may be suitable for this submission.
>
> *Editor instructions*
>
> The JOSS submission bot @editorialbot <https://github.com/editorialbot>
> is here to help you find and assign reviewers and start the main review. To
> find out what @editorialbot <https://github.com/editorialbot> can do for
> you type:
>
> @editorialbot commands
>
