### 2023-09-27 #5861 @arfon
@a91quaini – while we look for an editor to handle this submission, could you please check the DOIs [suggested above](https://github.com/openjournals/joss-reviews/issues/5861#issuecomment-1729199235), and if they look correct, add them to your BibTeX file?
