### 2023-09-28 #5861 @jbytecode
@a91quaini - Thank you for submitting your work to JOSS. I am the handling editor for this submission.

Firstly, I will send invitations to those whose papers have recently been published in JOSS and are similar with high similarity measures. If I fail to receive enough feedback, I will ask you to suggest suitable reviewers if so.

Thank you in advance.
