### 2023-09-28 #5861 @jbytecode
👋👋👋 Dear @ArdiaD, @kboudt, @jipegit, @mdsteiner, @zmjones, @GabrieleRovigatti, @LucyMcGowan 👋👋👋

**intrinsicFRP: an R Package for Efficient Estimation and Inference of Asset Pricing Models**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5861).

Thank you in advance!
