+++ paper/paper.bib
@@ -3,7 +3,8 @@
   author={Quaini, Alberto and Trojani, Fabio and Yuan, Ming},
   journal={Swiss Finance Institute Research Paper},
   number={23-81},
-  year={2023}
+  year={2023},
+  doi={10.2139/ssrn.4574683}
 }
 
 @article{kan2013pricing,
@@ -14,18 +15,8 @@
   number={6},
   pages={2617--2649},
   year={2013},
-  publisher={Wiley Online Library}
-}
-
-@article{kan2013pricing,
-  title={Pricing model performance and the two-pass cross-sectional regression methodology},
-  author={Kan, Raymond and Robotti, Cesare and Shanken, Jay},
-  journal={The Journal of Finance},
-  volume={68},
-  number={6},
-  pages={2617--2649},
-  year={2013},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.3386/w15047}
 }
 
 @article{fama1973risk,
@@ -36,7 +27,8 @@
   number={3},
   pages={607--636},
   year={1973},
-  publisher={The University of Chicago Press}
+  publisher={The University of Chicago Press},
+  doi={10.1086/260061}
 }
 
 @article{shanken1992estimation,
@@ -47,7 +39,8 @@
   number={1},
   pages={1--33},
   year={1992},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/rfs/5.1.1}
 }
 
 @article{kan1999two,
@@ -58,7 +51,8 @@
   number={1},
   pages={203--235},
   year={1999},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/0022-1082.00102}
 }
 
 @article{gospodinov2014misspecification,
@@ -69,7 +63,8 @@
   number={7},
   pages={2139--2170},
   year={2014},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.2139/ssrn.2579821}
 }
 
 @article{kleibergen2009tests,
@@ -80,7 +75,8 @@
   number={2},
   pages={149--173},
   year={2009},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jeconom.2009.01.013}
 }
 
 @article{hansen1997assessing,
@@ -91,7 +87,8 @@
   number={2},
   pages={557--590},
   year={1997},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.3386/t0153}
 }
 
 @article{chen2019improved,
@@ -102,7 +99,8 @@
   number={4},
   pages={1787--1824},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.2139/ssrn.3177681}
 }
 
 @article{kleibergen2006generalized,
@@ -113,7 +111,8 @@
   number={1},
   pages={97--126},
   year={2006},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.2139/ssrn.373420}
 }
 
 @article{sanderson2016armadillo,
@@ -123,7 +122,8 @@
   volume={1},
   number={2},
   pages={26},
-  year={2016}
+  year={2016},
+  doi={10.21105/joss.00026}
 }
 
 @article{eddelbuettel2018extending,
@@ -134,7 +134,8 @@
   number={1},
   pages={28--36},
   year={2018},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/00031305.2017.1375990}
 }
 
[…]
