+++ DESCRIPTION
@@ -12,7 +12,7 @@ Description: Tradable factor risk premia are given by the negative factor covari
     with the Stochastic Discount Factor projection on returns. This package 
     provides efficient computation of tradable and Oracle tradable factor risk 
     premia estimators and their standard errors; see A. Quaini, F. Trojani and 
-    M. Yuan (2023) <https://papers.ssrn.com/sol3/papers.cfm?abstract_id=4574683>. 
+    M. Yuan (2023) <doi:10.2139/ssrn.4574683>. 
     Tradable factor risk premia are robust to misspecification or weak 
     identification in asset pricing models, and they are zero for any factor 
     weakly correlated with returns. Their Oracle estimator performs as well as 
