### 2023-10-01 #5861 @a91quaini
Dear @jbytecode, thank you for your help.
I suggest the following potential reviewers, based on their knowledge of both R and C++, and, where possible, proximity in area of expertise:

stefanradev
mashrur-ayon
raybecker
AmirHoseinHadian
mmrabe

I have also added the doi's to the BibTex.
Thank you again!
