### 2023-10-01 #5861 @jbytecode
@a91quaini - Thank you for correcting the DOI issue. The list you provided is very helpful, indeed.
