### 2023-10-01 #5861 @mashrur-ayon
> 👋👋👋 Dear  @raybecker and @mashrur-ayon 👋👋👋
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled:
>
> **intrinsicFRP: an R Package for Efficient Estimation and Inference of Asset Pricing Models**
>
> You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5861).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be 23 check items for each single reviewer.
>
> Thank you in advance!

I am ready to take the review 😎 thanks for offering me.
