### 2023-10-05 #5861 @jbytecode
👋👋👋 Dear  @mmrabe 👋👋👋

**intrinsicFRP: an R Package for Efficient Estimation and Inference of Asset Pricing Models**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5861).

Thank you in advance!
