### 2023-10-09 #5861 @mmrabe
Dear @jbytecode,

Thank you for considering me as a reviewer. I'm afraid I don't have the necessary expertise in quantitative finance to review this submission and therefore unfortunately have to decline.
