### 2023-10-09 #5861 @raybecker
Sorry for the delay. Clearly I'm swamped at the moment and I fear that
financial models are not my expertise, so I'm afraid that I must decline.
Good luck finding another person to review.

On Sun, Oct 1, 2023 at 8:18 PM Mehmet Hakan Satman ***@***.***>
wrote:

> 👋👋👋 Dear @raybecker <https://github.com/raybecker> and @mashrur-ayon
> <https://github.com/mashrur-ayon> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled:
>
> *intrinsicFRP: an R Package for Efficient Estimation and Inference of
> Asset Pricing Models*
>
> You can find more information at the top of this Github issue (#5861
> <https://github.com/openjournals/joss-reviews/issues/5861>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a separate thread. In that thread, there will
> be 23 check items for each single reviewer.
>
> Thank you in advance!
>
