### 2023-10-10 #5861 @AmirHoseinHadian
Dear Mehmet Hakan,

Thanks for contacting me and considering me as the reviewer. But I am not
sure if I can be a good person for reviewing this paper. My expertise in
computational modeling in cognitive science and I'm not familiar with the
context of finance.
Then I should apologize and decline the review process. By the way I will
suggest you Dr. Jamal Amani Rad as the reviewer. He is expert in the
context of option pricing and also he is familiar with R and C++.
His Email contact is: ***@***.***
His scholar page:
https://scholar.google.com/citations?user=LA2VT8kAAAAJ&hl=en
His github profile: https://github.com/amanirad

Kind regards,
Amir

On Tue, 10 Oct 2023 at 07:14, Mehmet Hakan Satman ***@***.***>
wrote:

> @a91quaini <https://github.com/a91quaini> - It seems the suggested
> potential reviewers either don't response or are not available for this
> submission. Do you have additional set of potential reviewers? Otherwise,
> I'll try to find one for you. Thank you in advance.
>
