### 2023-10-12 #5861 @jbytecode
@GabrieleRovigatti - Thank you for the suggestion.

@LeopoldoCatania - You have been suggested as a potential reviewer for this submission. Would you like to assist in reviewing this submission for JOSS?
