### 2023-10-16 #5861 @jbytecode
👋👋👋 Dear  @mingzehuang and @lrnv 👋👋👋

**We need one more reviewer for this submission**

**intrinsicFRP: an R Package for Efficient Estimation and Inference of Asset Pricing Models**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5861).

Thank you in advance!
