### 2023-10-16 #5861 @lrnv
@jbytecode I'll gladly take this one if you still need someone, asset pricing is fine with me :)
