### 2023-10-16 #5954 @editorialbot
[REVIEW]: intrinsicFRP: an R Package for Efficient Estimation and Inference of Asset Pricing Models
**Submitting author:** @a91quaini (<a href="http://orcid.org/0000-0002-1251-0599">Alberto Quaini</a>)
**Repository:** https://github.com/a91quaini/intrinsicFRP
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
