### 2023-11-05 a91quaini/intrinsicFRP#2 @lrnv
[docs] Documentation need to contain references
For exemples, in the `ChenFang2019BetaRankTest` documentation, you state that is compute the "Chen fang 2019 rank statistic and p-value". I bet that Chen-Fang 2019 is a reference to a given paper. The documentation could include this reference in a better format (like in a paper) so that getting to the source is easier.

More generally, each page of the documentation that refers to the litterature should include the corresponding bibliography. I havent done that in a while, but I remember that there are tools to include properly-formated bibtex references in Roxygen documentation, maybe you could take a look ?

------------

Other instance: ?TRFP with Newey-West 1994

----------

Same thing with the datasets: there is currently a link but in the paper you say that there data come from (French, 2022) ? This ref needs to appear in the docs too
