### 2023-11-05 a91quaini/intrinsicFRP#3 @lrnv
[docs] Community guidelines
In at least the readme, you should include a block of text that indicates to someone that might need the info how they mlight 1) contribute tot he software, 2) report issues or problems, or 3° seek support. Usually one paragraph is enough.
