### 2023-11-05 a91quaini/intrinsicFRP#4 @lrnv
[readme] Bibliography is not full / not consistent
1) One reference ahs a link to the paper, the -two others do not : either you remove this link or you provide links for every reference.

2) Many papers are quoted in the readme without the references at the end, this is not OK: some people might not know the litterature as good as you :)
