### 2023-11-05 a91quaini/intrinsicFRP#5 @lrnv
[paper] Figure 1 too small
Make it as large as the text, it will be much better.
