### 2023-11-05 a91quaini/intrinsicFRP#6 @lrnv
[docs] Exemple on the real dataset
I bet you have somewhere an exemple of use of your package on the real datasets it contains. This exemple should appear :

1) In the readme, alongside the exemple that is already there.
2) In the documentation pages associated to the datasets, which for the moment have no exemples at all.

This will help people understand how usefull your code might be :)
