### 2023-11-05 a91quaini/intrinsicFRP#7 @lrnv
[docs] HJMisspecificationTest detaisl section seems abit harsh
You only state "Computes the Hansen-Jagannatan misspecification statistic and p-value of an asset pricing model
from test asset excess returns and risk factors.". This is a bit harsh as i it very hard to understand what is going on without already knowing what you are talking about.

COuld you developp on why this stat is important / how it can be used AND give the propper reference ?

-----

Same thing for ohter documentations entries. Generally, you hjave to expand as much as possible on each entry page. I know this is a bit of work, but I  assure you that it is worth it !
