### 2023-11-05 a91quaini/intrinsicFRP#8 @lrnv
[tests -> docs] Copy tests as documentation exemples
I see that your testing suite is very well organised and correspond to each of the functions that you are exporting. One good idea would be to use the exemples that you test your code onto as additional documentations exemples. This is a low hanging fruit but will be a great improvement to your docs
