### 2023-11-05 a91quaini/intrinsicFRP#9 @lrnv
[paper] Introduction : lackas a few more general sentences. 
The paper is very well writen and nice to read, thanks a lot for that !

Could you refactor a bit the first paragraph of the summary so that it is readable for people that are not from the fioannce field ? Think about people that have knowledge in computer science, R and Cpp but not in mathematics, probability, and finance: maybe from biology ?

For them, understanding what you are talking about might be very difficult in the current state of the first paragraphs. A good solution would be to include a few generic sentences introducing your field to non-specialist, including a good reference to a recent-but-novice-friendly book :)
