### 2023-11-05 a91quaini/intrinsicFRP#10 @lrnv
[paper] second paragraph source ?
The second paragraph in itself is a strong result, however, I do not see the source of it ? Include refernce please.
