### 2023-11-05 a91quaini/intrinsicFRP#11 @lrnv
[paper] statement of need precision 
You say that "By addressing the two issues [...] the paper allows to readily apply methods from (Qaini et al)". But is it the only methods it allows to apply ? I thought that the methods you documented and exported were comming from a broader piece of litterature?

Maybe you should reformulate a bit this sentence.
