### 2023-11-05 #5954 @lrnv
@a91quaini, @jbytecode,

I have taken the time to do the review, test the code myself and read the paper. Overall, I am very pleased with the paper, the package, and I can appreciate the effort and the amount of work that has been put into this submission. This is a very good work that definitely has its place in JOSS.

I have opened a few issues on the package repo :

- https://github.com/a91quaini/intrinsicFRP/issues/2
- https://github.com/a91quaini/intrinsicFRP/issues/3
- https://github.com/a91quaini/intrinsicFRP/issues/4
- https://github.com/a91quaini/intrinsicFRP/issues/5
- https://github.com/a91quaini/intrinsicFRP/issues/6
- https://github.com/a91quaini/intrinsicFRP/issues/7
- https://github.com/a91quaini/intrinsicFRP/issues/8
- https://github.com/a91quaini/intrinsicFRP/issues/9
- https://github.com/a91quaini/intrinsicFRP/issues/10
- https://github.com/a91quaini/intrinsicFRP/issues/11

Nothing really fundamental. The biggest issue for me is that the paper and the documentations are a bit lacking in terms of references. Since most of the implemented methods are known estimators / routines / stuff from the literature, the corresponding literature should appear everywhere in the docs.

Moreover, the test suite is very satisfactory, and one of my remarks is that the examples used in the test suite should also be given in the documentation as, well, examples.

An example on real data in the README and the docs would be very nice bonus to understand more clearly what the package does, since it also provide the real datasets.

Finally, the details given in the documentations are sometimes a bit harsh and hard to follow for non-specialists, and might be rewritten with more details in certain places. The full documentation is not very long (the reference manual on CRAN is only 10 pages, with only 7 entries), and thus the `Description` piece of each entry could benefit from a careful examination.

None of that should be very hard to fix, and I will reconsider the few unchecked boxes of my review when it is done. In the meantime, @a91quaini we can discuss each of these problems in the corresponding issues.
