### 2023-11-05 #5954 @a91quaini
@lrnv <https://github.com/lrnv>, thank you for your detailed review.

I will improve the documentation and include detailed references where
appropriate.

Actually, the example provided in the README involves real data. I will
stress this more clearly and add an interpretation of the results.
Thank you again.

Best, Alberto
