### 2023-11-19 f0a0577
huge commit: add hac covariance and correct misspecification test
