+++ src/tfrp.cpp
@@ -1,7 +1,7 @@
 // Author: Alberto Quaini
 
 #include "tfrp.h"
-#include "hac_standard_errors.h"
+#include "hac_covariance.h"
 
 //////////////////
 ///// TFRPCpp ////
@@ -12,12 +12,17 @@ Rcpp::List TFRPCpp(
   const bool include_standard_errors
 ) {
 
+  // Check if standard errors should be included in the output
   if (include_standard_errors) {
 
+    // Calculate the covariance between factors and returns
     const arma::mat covariance_factors_returns = arma::cov(factors, returns);
+    // Calculate the variance of the returns
     const arma::mat variance_returns = arma::cov(returns);
+    // Calculate the mean of the returns
     const arma::vec mean_returns = arma::mean(returns).t();
 
+    // Create a list containing both risk premia and standard errors
     return Rcpp::List::create(
       Rcpp::Named("risk_premia") = TFRPCpp(
         covariance_factors_returns,
@@ -35,6 +40,7 @@ Rcpp::List TFRPCpp(
 
   } else {
 
+    // If standard errors are not included, only return the risk premia
     return Rcpp::List::create(
       Rcpp::Named("risk_premia") = TFRPCpp(
         arma::cov(factors, returns),
@@ -42,7 +48,6 @@ Rcpp::List TFRPCpp(
         arma::mean(returns).t()
       )
     );
-
   }
 
 }
@@ -53,6 +58,7 @@ arma::vec TFRPCpp(
   const arma::vec& mean_returns
 ) {
 
+  // Return the risk premia computed using the formula Cov(F, R) * Var(R)^(-1) * E(R)
   return covariance_factors_returns * arma::solve(
     variance_returns,
     mean_returns,
@@ -72,25 +78,22 @@ arma::vec StandardErrorsTFRPCpp(
   const arma::vec& mean_returns
 ) {
 
-  const arma::mat var_ret_inv_cov_ret_fac = arma::solve(
-    variance_returns,
-    covariance_factors_returns.t(),
-    arma::solve_opts::likely_sympd
-  );
+  // Compute the inverse of variance returns applied to two different matrices
+  const arma::mat var_ret_inv = arma::inv_sympd(variance_returns);
+  const arma::mat var_ret_inv_cov_ret_fac = var_ret_inv *
+    covariance_factors_returns.t();
 
+  // Center the returns and factors by subtracting their means
   const arma::mat returns_centred = returns.each_row() - mean_returns.t();
-
   const arma::mat factors_centred = factors.each_row() - arma::mean(factors);
 
-  const arma::vec ret_cen_var_ret_inv_mean_ret = returns_centred * arma::solve(
-    variance_returns,
-    mean_returns,
-    arma::solve_opts::likely_sympd
-  );
-
+  // Compute intermediate terms used in the standard error computation
+  const arma::vec ret_cen_var_ret_inv_mean_ret = returns_centred *
+    var_ret_inv * mean_returns;
   const arma::mat ret_cen_var_ret_inv_cov_ret_fac = returns_centred *
     var_ret_inv_cov_ret_fac;
 
+  // Calculate the standard errors using HACStandardErrorsCpp function
   return HACStandardErrorsCpp(
     // covariance term
     factors_centred.each_col() % ret_cen_var_ret_inv_mean_ret -
@@ -99,6 +102,6 @@ arma::vec StandardErrorsTFRPCpp(
     ret_cen_var_ret_inv_mean_ret +
     // mean term
     returns_centred * var_ret_inv_cov_ret_fac
-  );
+  ) / std::sqrt(static_cast<double>(returns.n_rows));
 
 }
