### 2023-11-19 a91quaini/intrinsicFRP#2 @a91quaini
Thank you, I have updated the documentation of the functions. As for the dataset, the link to the Kenneth French data library can be found in the appropriate dataset doc, e.g., by running ?intrinsicFRP::returns. I removed the paper (French, 2022) as it was just a link to the library.
