### 2023-11-20 e2393f2
test cpp arguments checks
