### 2023-11-20 #5954 @a91quaini
Dear @jbytecode <https://github.com/jbytecode>, @lrnv
<https://github.com/lrnv> and @mashrur-ayon
<https://github.com/mashrur-ayon>.

I wanted to thank you for your time so far, especially @jbytecode
<https://github.com/jbytecode> for finding the reviewers and @lrnv
<https://github.com/lrnv> for the useful comments.

In view of the comments of @lrnv <https://github.com/lrnv> and in view of
my recent research projects that require considerable expansions of the
methods implemented in the package, I would like to withdraw my submission
to the JOSS. I want to make an effort to transform this package into a
complete suit for estimating and testing factor models in asset pricing.
Given the amount of time that it already took me and the amount of time
that it will take me to further develop it, I wanted to aim for a journal
that will be counted for my tenure criteria (unfortunately, JOSS isn't on
the list).

I apologize for any inconvenience, I hope that the help you provided did
not take too much of your time and I hope you will understand my choice. I
have already informed @lrnv <https://github.com/lrnv> earlier today asking
if he was ok with my choice, given that he spent some time reviewing the
package.

Kind regards,
Alberto

On Sun, 19 Nov 2023 at 18:51, Mehmet Hakan Satman ***@***.***>
wrote:

> @a91quaini <https://github.com/a91quaini> - Could you please update your
> status and inform us on how is your work going? Thank you in advance.
>
